microsoft-mcp/requirements.txt
fastmcp>=2.8.0
httpx>=0.27.0
msal>=1.32.3
python-dotenv>=1.0.0
flask>=3.0.0
